mono/requirements.txt
tensorboard==1.10.0
tensorflow-gpu==1.10.0
numpy==1.14.5
tqdm==4.46.1
opencv-python==4.2.0.34
matplotlib==3.2.2